feat: add header with animation
