+++ src/pages/Home.tsx
@@ -4,17 +4,28 @@ import { useSnapshot } from 'valtio'
 
 import state from '../store'
 
-// import {
-//   headContainerAnimation,
-//   headContentAnimation,
-//   headTextAnimation,
-//   slideAnimation
-// } from '../config/motion'
+import {
+  Direction,
+  headContainerAnimation,
+  headContentAnimation,
+  headTextAnimation,
+  slideAnimation
+} from '../config/motion'
 
 const Home: React.FC = (): JSX.Element => {
   const snap = useSnapshot(state)
-  
-  return <div>Home</div>
+
+  return (
+    <AnimatePresence>
+      {snap.intro && (
+        <motion.div className="home" {...slideAnimation(Direction.LEFT)}>
+            <motion.header>
+              <img src="./threejs.png" alt="logo" className="w-8 h-8 object-contain" />
+            </motion.header>
+        </motion.div>
+      )}
+    </AnimatePresence>
+  )
 }
 
 export default Home
